Removed inheritance
Instead of having some UI element or bonuses inherit from the player class, which was causing all sorts of bugs. Well not bugs per say but it was weird and confusing. So instead all the element that needs to access the player data must have a reference to the Scriptable object "PlayerData".